python_merchant/requirements.txt
google-shopping-merchant-datasources
google-shopping-merchant-inventories
google-shopping-merchant-products
google-auth-oauthlib
google-auth
googleapis-common-protos